manage condition fro drop schema and add has permission in master blade

+++ src/Controllers/PackageController.php
@@ -75,7 +75,7 @@ class PackageController extends Controller
                         '--force' => true,
                         ]);
                     }
-            
+
                     if (is_dir(base_path('vendor/admin/admin_role_permissions'))) {
                         Artisan::call('db:seed', [
                             '--class' => 'Packages\\Admin\\AdminRolePermissions\\database\\seeders\\AssignAdminRoleSeeder',
@@ -98,7 +98,6 @@ class PackageController extends Controller
             }
 
             return back()->with('success', $message);
-
         } catch (\Exception $e) {
             if ($request->expectsJson()) {
                 return response()->json(['status' => 'error', 'message' => $e->getMessage()], 500);
@@ -128,19 +127,23 @@ class PackageController extends Controller
         }
 
         if (Schema::hasTable($package)) {
-            Schema::drop($package);
+            if ($package != 'admins' && Schema::hasTable('admins')) {
+                Schema::drop($package);
+            }
         }
 
         // If package is 'users', also drop user_roles table
         if ($package === 'users' && Schema::hasTable('user_roles')) {
             Schema::drop('user_roles');
         }
-        
-        \DB::table('migrations')
-          ->where('migration', 'like', '%create_'.$package.'_table%')
-          ->delete();
 
-         // Also remove user_roles migration record if applicable
+        if ($package != 'admins') {
+            \DB::table('migrations')
+                ->where('migration', 'like', '%create_' . $package . '_table%')
+                ->delete();
+        }
+
+        // Also remove user_roles migration record if applicable
         if ($package === 'users') {
             \DB::table('migrations')
                 ->where('migration', 'like', '%create_user_roles_table%')
